backend/requirements.txt
# Django Core
Django>=4.2.0,<5.0
djangorestframework>=3.14.0

# Authentication & Security
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0

# File Handling & Storage
Pillow>=10.0.0
python-magic>=0.4.27

# Encryption & Security
cryptography>=41.0.0
bcrypt>=4.0.0

# Payment Processing
stripe>=5.5.0

# Environment Management
python-decouple>=3.8
django-environ>=0.10.0

# Development Tools (optional)
django-debug-toolbar>=4.1.0
django-extensions>=3.2.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.2.0

# Production Dependencies
gunicorn>=21.0.0
whitenoise>=6.5.0

# Database (for production)
psycopg2-binary>=2.9.0

# Task Queue (for background processing)
celery>=5.3.0
redis>=4.6.0